v3.5.49 Fix undefined daily goal for Sundays (#33)
When is Sunday, an anomaly occurs caused by difference between the
way JavaScript Date API controls the days (starting day is Sunday
and ending day is Saturday) and Kairos control mode (starting day
is Monday and ending day is Sunday).

The settings have 0 (zero) for identify a Sunday. When a variable
uses 7 to control loops and decisions, an error occurs because 7
does not exist in the settings. The big decision here is how Sundays
should be controlled: with 0 or with 7.

This commit fixes this behavior, the difference between two ways is
managed correctly, doing the parse when necessary.